Challenge completed
This tookk a lot longer than I would have liked, the restriction of only
being allowed to write a single line with ternary operators.

This was quite frustrating for me since I knew what I needed to do but
would have preferred to write it as several lines of code.